requirements.txt
boto3==1.15.11
numpy==1.19.2
pandas==1.1.2
psycopg2==2.8.6
PyYAML==5.3.1
testing.postgresql==1.3.0